feat(galacy-gallery): prepare for join ux
